requirements.txt
args==0.1.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
clint==0.5.1
cloudinary==1.8.0
cryptography==1.7.2
dj-database-url==0.4.2
Django==1.11.5
django-dotenv==1.4.1
future==0.16.0
gunicorn==19.7.1
idna==2.6
jwt==0.5.2
line-bot-sdk==1.3.0
mock==2.0.0
pbr==3.1.1
psycopg2==2.7.3.1
pyasn1==0.3.4
pycparser==2.18
PyJWT==1.5.3
pytz==2017.2
requests==2.18.4
requests-toolbelt==0.8.0
six==1.10.0
urllib3==1.22
whitenoise==3.3.0
